api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
qdrant-client==1.12.1
sentence-transformers==3.0.1
openai==1.52.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
numpy==1.26.4
asyncio==3.4.3
python-multipart==0.0.20